producer/requirements.txt
kafka-python
numpy